Readds the openat seccomp rule that was replaced on #25440. #23715
While the fix for #25440 worked on some systems, on others the old
rule was actually working and #25440 broke tor for them. To make
tor work on both kinds of systems, both rules are now added
because it is not known what causes one rule to work on some
systems and other on some other systems.